+++ src/client-graphql/modules/default.ts
@@ -1,18 +1,9 @@
-import {
-  GraphQLObjectType,
-  GraphQLFieldConfigMap,
-  GraphQLNonNull,
-  GraphQLList,
-  GraphQLScalarType,
-  GraphQLFloat,
-  GraphQLString,
-  GraphQLResolveInfo,
-} from "graphql";
+import { GraphQLObjectType, GraphQLFieldConfigMap, GraphQLNonNull, GraphQLList, GraphQLResolveInfo } from "graphql";
 import { TableByName, ModuleFieldResolverParent } from "./module-plugin";
 import { getUniqueTypeName, toSafeName } from "../shared-functions";
-import { ProductTableFileColumn, ProductTableFileColumnType } from "../..";
 import { Context } from "../context";
 import { TableRow } from "../schema-types";
+import { resolveTable, buildTableRowTypeFields } from "./shared-functions";
 
 /**
  * This is the default generic handling for modules
@@ -52,63 +43,11 @@ function buildModuleType(moduleName: string, tableByName: TableByName, usedTypeN
   return new GraphQLObjectType({ name: getUniqueTypeName(`Module_${moduleName}`, usedTypeNames), fields });
 }
 
-export function buildTableRowTypeFields(
-  columns: ReadonlyArray<ProductTableFileColumn>
-): GraphQLFieldConfigMap<unknown, unknown> {
-  return Object.fromEntries(
-    columns.map((c) => [toSafeName(c.name), { type: columnTypeToGraphQLType(c), description: c.description }])
-  );
-}
-
-function columnTypeToGraphQLType(c: ProductTableFileColumn): GraphQLScalarType {
-  switch (c.type) {
-    case ProductTableFileColumnType.Number:
-      return GraphQLFloat;
-    case ProductTableFileColumnType.Blob:
-    case ProductTableFileColumnType.DynamicDiscrete:
-    case ProductTableFileColumnType.FixedDiscrete:
-    case ProductTableFileColumnType.ForeignKey:
-    case ProductTableFileColumnType.LongText:
-    case ProductTableFileColumnType.PrimaryKey:
-    case ProductTableFileColumnType.Product:
-    case ProductTableFileColumnType.Property:
-    case ProductTableFileColumnType.PropertyFilter:
-    case ProductTableFileColumnType.PropertyValues:
-    case ProductTableFileColumnType.Quantity:
-    case ProductTableFileColumnType.Text:
-    case ProductTableFileColumnType.TextId:
-    case ProductTableFileColumnType.Unit:
-      return GraphQLString;
-    default:
-      return GraphQLString;
-  }
-}
-
-export async function moduleTableResolver(
+async function moduleTableResolver(
   parent: ModuleFieldResolverParent,
   _args: {},
   ctx: Context,
   info: GraphQLResolveInfo
 ): Promise<ReadonlyArray<TableRow>> {
-  return resolveTable(parent.module, parent.productFileName, info.fieldName, ctx);
-}
-
-export async function resolveTable(
-  module: string,
-  productFileName: string,
-  tableName: string,
-  ctx: Context
-): Promise<ReadonlyArray<TableRow>> {
-  const fullTableName = `${module}@${tableName}`;
-  const productFile = await ctx.loaders.productFiles.load(productFileName);
-  const tableRef = productFile.data.tables[fullTableName];
-  const tableFileName = productFile.refs[tableRef];
-  if (!tableFileName) {
-    return [];
-  }
-  const tableFile = await ctx.loaders.tableFiles.load(tableFileName);
-  const rows = tableFile.data.rows.map((values) =>
-    Object.fromEntries(tableFile.data.columns.map((c, i) => [c.name, values[i]]))
-  );
-  return rows;
+  return resolveTable(parent.module, parent.productFileName, info.fieldName, ctx.loaders);
 }
