+++ src/client-graphql/modules/properties.ts
@@ -1,8 +1,12 @@
 import { GraphQLObjectType, GraphQLNonNull, GraphQLFieldConfigMap, GraphQLList } from "graphql";
 import { getUniqueTypeName } from "../shared-functions";
 import { TableByName, ModuleFieldResolverParent } from "./module-plugin";
-import { buildTableRowTypeFields, resolveTable } from "./default";
 import { Context } from "../context";
+import {
+  resolveTable,
+  buildTableRowTypeFields,
+  getProductFileNameFromRow as getProductFileName,
+} from "./shared-functions";
 
 /**
  * This file has specific schema and resolvers for the properties module
@@ -27,9 +31,8 @@ export async function createModuleType(
       values: {
         type: new GraphQLNonNull(new GraphQLList(new GraphQLNonNull(propertyValueRowType))),
         resolve: async (parent, _args, ctx) => {
-          console.log("Resolve values...", parent);
-          const rows = await resolveTable("properties", parent._productFileName, "property.value", ctx);
-          return rows;
+          const rows = await resolveTable("properties", getProductFileName(parent), "property.value", ctx.loaders);
+          return rows.filter((r) => r["builtin@parent_id"] === parent["builtin@id"]);
         },
       },
     },
@@ -37,39 +40,10 @@ export async function createModuleType(
   fields["property"] = {
     type: new GraphQLNonNull(GraphQLList(new GraphQLNonNull(propertyTableRowType))),
     description: propertyTable.description,
-    // resolve: moduleTableResolver,
-    resolve: async (parent: ModuleFieldResolverParent, _args, ctx: Context) => {
-      console.log("Resolve property", parent);
-      const rows = await resolveTable(parent.module, parent.productFileName, "property", ctx);
-      // tslint:disable-next-line:no-any
-      for (const r of rows) {
-        // tslint:disable-next-line:no-any
-        (r as any)._productFileName = parent.productFileName;
-      }
-      return rows;
-    },
+    resolve: async (parent: ModuleFieldResolverParent, _args, ctx: Context) =>
+      resolveTable(parent.module, parent.productFileName, "property", ctx.loaders, true),
   };
   return new GraphQLObjectType({ name: getUniqueTypeName(`Module_${moduleName}`, usedTypeNames), fields });
 }
 
 export { resolveModuleType } from "./default";
-
-// export async function propertyTableResolver(
-//   parent: ModuleFieldResolverParent,
-//   args: {},
-//   ctx: Context,
-//   info: GraphQLResolveInfo
-// ): Promise<ReadonlyArray<TableRow>> {
-//   const rows = await moduleTableResolver(parent, args, ctx, info);
-//   return rows;
-// }
-
-// export async function propertyRowResolver(
-//   parent: ProductTableFileRow,
-//   args: {},
-//   ctx: Context,
-//   info: GraphQLResolveInfo
-// ): Promise<ReadonlyArray<TableRow>> {
-//   const rows = await moduleTableResolver(parent, args, ctx, info);
-//   return rows;
-// }
